Add togglable deceleration (want to refactor later)
Currently deceleration is a function that reduces the float value of the
x velocity in the Vector2 velocity which is added to the position. the
x velocity can't be reduced to a value less than 0. In the future I want
to refactor deceleration by creating a temp var to hold a deceleration
value which is then added to current velocity. Then that value is added
to the position.