Add rotation ui and homing bullet prototype
Added ability to control rotation speed and direction via ui controls.
Also worked on a homing bullet prototype that will be able to home onto
a moving target such as a player. Next will be to implement this new
bullet subclass into the ui so that the homing agressiveness can be
manipulated and a player instance is added allowing control over said
player to demonstrate the homing feature.