Add primative bullet children spawn feature
When a linear bullet is instantiated into the world it will be able to
create a corresponding child bullet with it's own pre-defined
properties. This is accomplished by creating a linear_bullet_factory to
then create an new child bullet instance when a singal is emitted from
the parent bullet.